Bug 1865694 - Update libvpx to f6b7166a2b6bac544c2c487d3a7e49bc265cdf9d r=webrtc-reviewers,pehrsons
This patch updates libvpx by two steps:

1. ./mach vendor media/libvpx/moz.yaml --patch-mode=none
2. ./mach vendor media/libvpx/moz.yaml --patch-mode=only

The first step pulls the upstream repo files to gecko, and the second
step applies the local patches.

Differential Revision: https://phabricator.services.mozilla.com/D197546

UltraBlame original commit: 7048555a140d845c2ddb403ede3045878ceccafa

+++ media/libvpx/libvpx/vp8/encoder/ratectrl.c
@@ -719,7 +719,8 @@ static void calc_pframe_target_size(VP8_COMP *cpi) {
         }
 
         
-        cpi->this_frame_target -= (cpi->this_frame_target * percent_low) / 200;
+        cpi->this_frame_target -=
+            (int)(((int64_t)cpi->this_frame_target * percent_low) / 200);
